refactor selectors (#265)
* refactor selectors

* move src/utils/selectors -> src/selectors
