+++ src/components/CreateNicRow/tests/CreateNicRow.test.js
@@ -9,7 +9,7 @@ import { Dropdown } from '../../Form/Dropdown';
 import CreateNicRowFixture from '../fixtures/CreateNicRow.fixture';
 import { networkConfigs } from '../../../tests/mocks/networkAttachmentDefinition';
 import { cloudInitTestVm } from '../../../tests/mocks/vm/cloudInitTestVm.mock';
-import { getName } from '../../../utils';
+import { getName } from '../../../selectors';
 import { Loading } from '../../Loading';
 import { POD_NETWORK } from '../../../constants';
