+++ src/components/Details/Flavor/fixtures/Flavor.fixture.js
@@ -2,7 +2,7 @@ import { Flavor } from '..';
 
 import { cloudInitTestVm } from '../../../../tests/mocks/vm/cloudInitTestVm.mock';
 import { fedora28 } from '../../../../k8s/objects/template/fedora28';
-import { getFlavor } from '../../../../utils';
+import { getFlavor } from '../../../../selectors';
 
 export default [
   {
