+++ src/selectors/host/selectors.js
@@ -1,0 +1,3 @@
+import { get } from 'lodash';
+
+export const getHostStatus = host => get(host, ['metadata', 'labels', 'metalkube.org/operational-status']);
