+++ src/selectors/index.js
@@ -1,0 +1,7 @@
+export * from './common';
+export * from './dv';
+export * from './host';
+export * from './pod';
+export * from './pvc';
+export * from './vm';
+export * from './vmi';
