+++ src/tests/forms_mocks/nic.mock.js
@@ -1,7 +1,7 @@
 import { NETWORK_TYPE_POD, NETWORK_TYPE_MULTUS } from '../../components/Wizard/CreateVmWizard/constants';
 import { POD_NETWORK } from '../../constants';
 import { network1 } from '../mocks/networkAttachmentDefinition/network.mock';
-import { getName } from '../../utils';
+import { getName } from '../../selectors';
 
 export const podNetwork = {
   name: 'podNetworkName',
