+++ src/utils/index.js
@@ -1,5 +1,4 @@
 export * from './mediaQuery';
-export * from './selectors';
 export * from './templates';
 export * from './utils';
 export * from './validations';
